requirements.txt
asgiref==3.7.2
Django==3.2.20
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.3
pkg-resources==0.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==0.21.1
pytz==2023.3
serializers==0.2.4
sqlparse==0.4.4
typing-extensions==4.7.1
